+++ packages/components/src/styles/checkbox-radio-styles.ts
@@ -121,12 +121,14 @@ export const getCheckboxRadioJssStyle = (
           ...(!isHighContrastMode &&
             hoverMediaQuery({
               '&(input:hover), .text:hover ~ &(input)': {
+                transition: 'unset', // Fixes chrome bug where border-color is stuck on hover color
                 borderColor: uncheckedHoverColor,
                 ...prefersColorSchemeDarkMediaQuery(theme, {
                   borderColor: uncheckedHoverColorDark,
                 }),
               },
               '&(input:checked:hover), .text:hover ~ &(input:checked)': {
+                transition: 'unset', // Fixes chrome bug where border-color is stuck on hover color
                 borderColor: checkedHoverColor,
                 backgroundColor: checkedHoverColor,
                 ...prefersColorSchemeDarkMediaQuery(theme, {